+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Create 2024.4.0-RC.3 build
+- REMIX-2783: Added light manipulator for CylinderLight
 
 ### Changed
 - Updated repo tools + added public version of repo lint